% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_unsd_sdg.R
\name{clean_sdg_series_data}
\alias{clean_sdg_series_data}
\title{Preprocess SDG indicators in line with the template.}
\usage{
clean_sdg_series_data(df.series)
}
\arguments{
\item{df.series}{A tibble of indicator data as returned by \code{combine_sdg_series_data}.}
}
\value{
A cleaned tibble of SDG indicators.
}
\description{
Preprocessess a combined dataframe of SDG indicators in line with
the template. In particular, renames the columns, removes redundant
disaggregation, streamlines data types and adds relevant columns. It
also runs sanity checks to validate the output.
}
\examples{
clean_sdg_series_data(df.series)
}
\seealso{
\code{\link[=combine_sdg_series_data]{combine_sdg_series_data()}} that combines the indicators.
}
