% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_unsd_sdg.R
\name{combine_sdg_series_data}
\alias{combine_sdg_series_data}
\title{Combine SDG indicators from a multi-sheet Excel file}
\usage{
combine_sdg_series_data(filepath, verbose = TRUE)
}
\arguments{
\item{filepath}{A path to the file produced by download_sgd_series_data.}
}
\value{
A combined tibble of SDG indicators.
}
\description{
This function combines SDG indicators from a multi-sheet Excel file into a
single tibble. In so doing, it select columns of interest, filters out
missing and invariant columns as well as renames some of those in line
with the agreed standard.
}
\examples{
combine_sdg_series_data("../data/tmp/unsd_indicators.xlsx")
}
\seealso{
\code{\link[=download_sgd_series_data]{download_sgd_series_data()}} that downloads the indicators.
}
