% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_unsd_sdgs.R
\name{exctract_sdg_series_list}
\alias{exctract_sdg_series_list}
\title{Extract a list of all available SDG indicators.}
\usage{
exctract_sdg_series_list()
}
\value{
A list of indicators and their metadata converted into a tibble.
}
\description{
This function is used to retrieve a list of all indicators available via UNSD SDGs API
in a tabular format.
}
\examples{
exctract_series_list()
}
