% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_unsd_sdgs.R
\name{get_sdg_api}
\alias{get_sdg_api}
\title{Make a GET request to UNSD SDG API}
\usage{
get_sdg_api(command)
}
\arguments{
\item{command}{An API request to submit.}
}
\value{
Contents of the response (parsed).
}
\description{
A generic function to make a GET request to United Nations Statistics Division SDG API.
See the \href{https://unstats.un.org/sdgapi/swagger/}{SDGs API documentation} for a
list of available commands.
}
\examples{
get_sdg_api("/v1/sdg/Indicator/List")
get_sdg_api("/v1/sdg/Goal/List")
}
\seealso{
\code{\link[=post_sdg_api]{post_sdg_api()}} which makes a POST request and accepts a \code{body} parameter.
}
