% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_unsd_sdgs.R
\name{post_sdg_api}
\alias{post_sdg_api}
\title{Make a POST request to UNSD SDG API}
\usage{
post_sdg_api(command, body)
}
\arguments{
\item{command}{An API request to submit.}

\item{body}{A body of the POST request.}
}
\value{
Contents of the parsed response converted into a tibble.
}
\description{
A generic function to make a POST request to United Nations Statistics Division SDG API.
See the \href{https://unstats.un.org/sdgapi/swagger/}{SDGs API documentation} for a
list of available commands.
}
\examples{
post_sdg_api("/v1/sdg/Series/DataCSV", list(seriesCodes = c("VC_DSR_MORT"), areaCodes = c(528)))
}
\seealso{
\code{\link[=get_sdg_api]{get_sdg_api()}} which makes a GET request returnes a response (parsed).
}
