% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_unsd_sdg.R
\name{validate_output}
\alias{validate_output}
\title{Validate output dataframe using sanity checks.}
\usage{
validate_output(df.output)
}
\arguments{
\item{df.output}{A tibble of indicator data within \code{clean_sdg_series_data}.}
}
\value{
Nothing
}
\description{
An internal function to validate the data produces within \code{clean_sdg_series_data}.
Currently only checks if duplicates result from data processing. If so,
issues a warning.
}
\seealso{
\code{\link[=clean_sdg_series_data]{clean_sdg_series_data()}} that calls this function internally.
}
